now are ALL packages lowercase
