--- src/main/java/mops/domain/models/Repository/QuestionPollRepositoryInterface.java
@@ -1,8 +1,0 @@
-package mops.domain.models.Repository;
-
-import mops.domain.models.QuestionPoll.QuestionPoll;
-import mops.domain.models.QuestionPoll.QuestionPollLink;
-
-public interface QuestionPollRepositoryInterface {
-  QuestionPollLink save(QuestionPoll questionPoll);
-}
