--- src/main/java/mops/domain/models/Repository/UserRepositoryInterface.java
@@ -1,10 +1,0 @@
-package mops.domain.models.Repository;
-
-import mops.domain.models.User.User;
-import mops.domain.models.User.UserId;
-
-public interface UserRepositoryInterface {
-
-  User getById(UserId id);
-
-}
