--- src/main/java/mops/domain/models/User/UserId.java
@@ -1,5 +1,0 @@
-package mops.domain.models.User;
-
-public class UserId {
-
-}
