+++ src/main/java/mops/domain/models/repository/QuestionPollRepositoryInterface.java
@@ -1,0 +1,8 @@
+package mops.domain.models.repository;
+
+import mops.domain.models.questionpoll.QuestionPoll;
+import mops.domain.models.questionpoll.QuestionPollLink;
+
+public interface QuestionPollRepositoryInterface {
+  QuestionPollLink save(QuestionPoll questionPoll);
+}
