+++ src/main/java/mops/domain/models/repository/UserRepositoryInterface.java
@@ -1,0 +1,10 @@
+package mops.domain.models.repository;
+
+import mops.domain.models.user.User;
+import mops.domain.models.user.UserId;
+
+public interface UserRepositoryInterface {
+
+  User getById(UserId id);
+
+}
