+++ src/main/java/mops/domain/models/user/User.java
@@ -1,4 +1,4 @@
-package mops.domain.models.User;
+package mops.domain.models.user;
 
 import lombok.Getter;
