+++ src/main/java/mops/domain/models/user/UserId.java
@@ -1,0 +1,5 @@
+package mops.domain.models.user;
+
+public class UserId {
+
+}
