+++ src/test/java/mops/domain/models/questionpoll/QuestionPollFactoryTest.java
@@ -1,7 +1,7 @@
-package mops.domain.models.QuestionPoll;
+package mops.domain.models.questionpoll;
 
 import mops.controller.dto.*;
-import mops.domain.models.User.UserId;
+import mops.domain.models.user.UserId;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
